refactor(metadata): add createImageMetadataFile function with logging for segmentation; improve error handling and directory creation
